+++ src/sequelize/index.js
@@ -41,6 +41,41 @@ function registerType(Sequelize) {
   PgTypes.VECTOR.parse = DataTypes.VECTOR.parse;
   PgTypes.VECTOR.types = {postgres: ['vector']};
   DataTypes.postgres.VECTOR.key = 'vector';
+
+  class HALFVEC extends ABSTRACT {
+    constructor(dimensions) {
+      super();
+      this._dimensions = dimensions;
+    }
+
+    toSql() {
+      return utils.halfvecType(this._dimensions).toUpperCase();
+    }
+
+    _stringify(value) {
+      return utils.toSql(value);
+    }
+
+    static parse(value) {
+      return utils.fromSql(value);
+    }
+  }
+
+  HALFVEC.prototype.key = HALFVEC.key = 'halfvec';
+
+  DataTypes.HALFVEC = Sequelize.Utils.classToInvokable(HALFVEC);
+  DataTypes.HALFVEC.types.postgres = ['halfvec'];
+
+  PgTypes.HALFVEC = function HALFVEC() {
+    if (!(this instanceof PgTypes.HALFVEC)) {
+      return new PgTypes.HALFVEC();
+    }
+    DataTypes.HALFVEC.apply(this, arguments);
+  };
+  util.inherits(PgTypes.HALFVEC, DataTypes.HALFVEC);
+  PgTypes.HALFVEC.parse = DataTypes.HALFVEC.parse;
+  PgTypes.HALFVEC.types = {postgres: ['halfvec']};
+  DataTypes.postgres.HALFVEC.key = 'halfvec';
 }
 
 function distance(op, column, value, sequelize, binary) {